removed util file
