+++ sdk/rust/src/lib.rs
@@ -13,7 +13,6 @@ use std::path::PathBuf;
 
 pub use porcupine;
 pub use rhino;
-pub mod util;
 
 pub struct PicovoiceBuilder<W, I>
 where
