+++ sdk/rust/tests/picovoice_tests.rs
@@ -9,6 +9,79 @@
     specific language governing permissions and limitations under the License.
 */
 
+mod platform {
+    #[allow(dead_code)]
+    const RPI_MACHINES: [&str; 4] = ["arm11", "cortex-a7", "cortex-a53", "cortex-a72"];
+    #[allow(dead_code)]
+    const JETSON_MACHINES: [&str; 1] = ["cortex-a57"];
+    
+    #[cfg(all(target_os = "linux", any(target_arch = "arm", target_arch = "aarch64")))]
+    fn find_machine_type() -> String {
+        use std::process::Command;
+    
+        let cpu_info = Command::new("cat")
+            .arg("/proc/cpuinfo")
+            .output()
+            .expect("Failed to retrieve cpu info");
+        let cpu_part_list = std::str::from_utf8(&cpu_info.stdout)
+            .unwrap()
+            .split("\n")
+            .filter(|x| x.contains("CPU part"))
+            .collect::<Vec<_>>();
+    
+        if cpu_part_list.len() == 0 {
+            panic!("Unsupported CPU");
+        }
+    
+        let cpu_part = cpu_part_list[0]
+            .split(" ")
+            .collect::<Vec<_>>()
+            .pop()
+            .unwrap()
+            .to_lowercase();
+    
+        let machine = match cpu_part.as_str() {
+            "0xb76" => "arm11",
+            "0xc07" => "cortex-a7",
+            "0xd03" => "cortex-a53",
+            "0xd07" => "cortex-a57",
+            "0xd08" => "cortex-a72",
+            "0xc08" => "beaglebone",
+            _ => "unsupported",
+        };
+    
+        String::from(machine)
+    }
+    
+    #[cfg(target_os = "macos")]
+    pub fn pv_platform() -> String {
+        String::from("mac")
+    }
+    
+    #[cfg(target_os = "windows")]
+    pub fn pv_platform() -> String {
+        String::from("windows")
+    }
+    
+    #[cfg(all(target_os = "linux", target_arch = "x86_64"))]
+    pub fn pv_platform() -> String {
+        String::from("linux")
+    }
+    
+    #[cfg(all(target_os = "linux", any(target_arch = "arm", target_arch = "aarch64")))]
+    pub fn pv_platform() -> String {
+        let machine = find_machine_type();
+        match machine.as_str() {
+            machine if RPI_MACHINES.contains(&machine) => String::from("raspberry-pi"),
+            machine if JETSON_MACHINES.contains(&machine) => String::from("jetson"),
+            "beaglebone" => String::from("beaglebone"),
+            _ => {
+                panic!("ERROR: Please be advised that this device is not officially supported by Picovoice");
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use std::collections::HashMap;
@@ -19,7 +92,7 @@ mod tests {
     use std::io::Read;
     use std::sync::{Arc, Mutex};
 
-    use picovoice::util::pv_platform;
+    use super::platform::pv_platform;
     use picovoice::{rhino::RhinoInference, Picovoice, PicovoiceBuilder};
 
     fn append_lang(path: &str, language: &str) -> String {
